[Chore] #47 - 코드리뷰 반영
