+++ Memento-iOS/Memento-iOS/Source/Extension/Date+.swift
@@ -100,4 +100,13 @@ extension Date {
         }
         return Date() > date
     }
+    
+    /// 셋팅뷰에서 wakeUpTime / sleepTime 등을 표시할 때 사용
+    /// 현재 Date 값을 "HH:mm" 형식(24시간제 시:분) 문자열로 반환
+    func formattedHourMinute() -> String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "HH:mm"
+        formatter.timeZone = TimeZone.current
+        return formatter.string(from: self)
+    }
 }
