+++ Memento-iOS/Memento-iOS/Source/Network/Setting/UserInfoUpdateAPIService.swift
@@ -6,8 +6,6 @@
 //
 
 import Foundation
-
-import Foundation
 import Moya
 
 // MARK: - UserInfoUpdateAPIServiceProtocol
